generalize pre/postprocessing
