+++ skellytracker/tests/sidecar/test_runtime.py
@@ -1,5 +1,9 @@
-"""Tests for skellytracker.core.sidecar.runtime — normalization dispatch and
-OnnxModelSpec construction from a resolved sidecar artifact.
+"""Tests for skellytracker.core.sidecar.runtime — OnnxModelSpec construction
+from a resolved sidecar artifact.
+
+Normalization dispatch (`build_normalization_fn`/`resolve_normalization_mode`)
+moved to `skellytracker.core.detectors.image_preprocessing` — see
+`tests/test_image_preprocessing.py`.
 """
 
 from __future__ import annotations
@@ -7,16 +11,11 @@ from __future__ import annotations
 from pathlib import Path
 from unittest.mock import patch
 
-import numpy as np
 import pytest
 
 from skellytracker.core.sessions.model_registry import resolve_model_path
 from skellytracker.core.sidecar.model import SidecarModel
-from skellytracker.core.sidecar.runtime import (
-    build_normalization_fn,
-    resolve_normalization_mode,
-    sidecar_model_spec,
-)
+from skellytracker.core.sidecar.runtime import sidecar_model_spec
 
 
 def _detector_sidecar_dict(**input_overrides) -> dict:
@@ -60,58 +59,6 @@ def _detector_sidecar_dict(**input_overrides) -> dict:
             }
         },
     }
-
-
-class TestNormalizationDispatch:
-    def test_none_mode_passes_through(self):
-        sidecar = SidecarModel.model_validate(
-            _detector_sidecar_dict(normalization="none")
-        )
-        fn = build_normalization_fn(sidecar.input, "fp32")
-        img = np.full((2, 2, 3), 200, dtype=np.uint8)
-        np.testing.assert_allclose(fn(img), img.astype(np.float32))
-
-    def test_unit_float_mode_divides_by_255(self):
-        sidecar = SidecarModel.model_validate(
-            _detector_sidecar_dict(normalization="unit_float")
-        )
-        fn = build_normalization_fn(sidecar.input, "fp32")
-        img = np.full((1, 1, 3), 255, dtype=np.uint8)
-        np.testing.assert_allclose(fn(img), np.ones((1, 1, 3), dtype=np.float32))
-
-    def test_imagenet_bgr_mode_matches_known_constants(self):
-        sidecar = SidecarModel.model_validate(
-            _detector_sidecar_dict(normalization="imagenet_bgr")
-        )
-        fn = build_normalization_fn(sidecar.input, "fp32")
-        img = np.zeros((1, 1, 3), dtype=np.uint8)
-        expected = (0.0 - np.array([123.675, 116.28, 103.53])) / np.array(
-            [58.395, 57.12, 57.375]
-        )
-        np.testing.assert_allclose(fn(img)[0, 0], expected, rtol=1e-5)
-
-    def test_custom_mode(self):
-        data = _detector_sidecar_dict(
-            normalization={
-                "mode": "custom",
-                "scale": 0.5,
-                "mean": [1.0, 1.0, 1.0],
-                "std": [2.0, 2.0, 2.0],
-            }
-        )
-        sidecar = SidecarModel.model_validate(data)
-        fn = build_normalization_fn(sidecar.input, "fp32")
-        img = np.full((1, 1, 3), 10, dtype=np.uint8)
-        # (10 * 0.5 - 1) / 2 = 2.0
-        np.testing.assert_allclose(fn(img)[0, 0], [2.0, 2.0, 2.0])
-
-    def test_normalization_by_precision_overrides_top_level(self):
-        data = _detector_sidecar_dict(
-            normalization="imagenet_bgr", normalization_by_precision={"int8": "none"}
-        )
-        sidecar = SidecarModel.model_validate(data)
-        assert resolve_normalization_mode(sidecar.input, "int8") == "none"
-        assert resolve_normalization_mode(sidecar.input, "fp32") == "imagenet_bgr"
 
 
 def _detector_sidecar_dict_with_url(**input_overrides) -> dict:
